+++ docs/12_0_0/gettingstarted/whatsnew.md
@@ -9,6 +9,9 @@ This page contains a list of big features. Please check the GitHub issues for al
 * CommandLink
   * Now has a widget and can be enabled and disabled by its widget.
   * Added `disableOnAjax` attribute to disable the link during Ajax requests triggered by it.
+* SplitButton
+  * Now can be enabled and disabled by its widget.
+  * Added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it or its menu items.
 
 Look into [migration guide](https://primefaces.github.io/primefaces/12_0_0/#/../migrationguide/12_0_0) for more enhancements and changes.